Move IEndpointStateChangeSubscriber implementations from the common module to their respective persistence bundles.
